- fix agentIp parse from metadata - fix download method - fix movelevel method

+++ src/itmo/escience/dstorage/utils/responses/DownloadResponse.java
@@ -52,7 +52,7 @@ public class DownloadResponse extends Response {
                     agentUri=(String)json.get(key);
                     break;
                 case "LVL": 
-                    level=(int)json.get(key);
+                    level=((Long)json.get(key)).intValue();
                     break;    
             }
         }
